requirements.txt
matplotlib==3.9.1
nltk==3.7
numpy==1.26.4
pillow==10.4.0
torch==2.2.2+cpu
torchaudio==2.2.2+cpu
torchvision==0.17.2+cpu

